Move hooking-related globals into a new HookBuilder class
