--- extension/src/webcat/interfaces/base.ts
@@ -1,4 +1,0 @@
-export enum hooksType {
-  page,
-  content_script,
-}
